refactor: DebugClient.java
I renamed the interface name of requestKeyType into RequestKeyType for camel case. and it's subclass name too.
 The inheritance classes inside the Command interface and the RequestKeyType interface were separated into their own files.
