+++ app/src/main/java/com/particlesdevs/photoncamera/debugclient/RequestKeyFloat.java
@@ -1,0 +1,20 @@
+package com.particlesdevs.photoncamera.debugclient;
+
+import android.hardware.camera2.CaptureRequest;
+
+class RequestKeyFloat implements RequestKeyType {
+    private CaptureRequest.Builder builder;
+    private CaptureRequest.Key<?> key;
+    String value;
+
+    public RequestKeyFloat(CaptureRequest.Builder builder, CaptureRequest.Key<?> key, String value) {
+        this.builder = builder;
+        this.key = key;
+        this.value = value;
+    }
+
+    @Override
+    public void setKey() {
+        builder.set((CaptureRequest.Key<Float>) key, Float.parseFloat(value));
+    }
+}
